+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.3] - 2024-01-21
+
+- Added `split_reversed` shader
+- Improved interval so that the effect lasts between 0 and 1
+
 ## [0.0.2] - 2023-12-26
 
 - Added `split` shader and changed the previous shader to `smoke` shader